+++ packages/api/schema/mutation.js
@@ -940,6 +940,24 @@ export default [
         orderPaymentId: ID!
         transactionContext: JSON
       ): String!
+
+      prepareProductMediaUpload(
+        mediaName: String!
+        productId: ID!
+      ): MediaUploadTicket!
+      prepareAssortmentMediaUpload(
+        mediaName: String!
+        assortmentId: ID!
+      ): MediaUploadTicket!
+      prepareUserAvatarUpload(
+        mediaName: String!
+        userId: ID
+      ): MediaUploadTicket!
+      confirmMediaUpload(
+        mediaUploadTicketId: ID!
+        size: Int!
+        type: String!
+      ): Media!
     }
   `,
 ];