+++ packages/core-files-next/core-files-next-tests.js
@@ -1,0 +1,12 @@
+// Import Tinytest from the tinytest Meteor package.
+/* import { Tinytest } from "meteor/tinytest";
+
+// Import and rename a variable exported by core-files-next.js.
+import { name as packageName } from "meteor/unchained:core-files-next";
+
+// Write your tests here!
+// Here is an example.
+Tinytest.add('core-files-next - example', function (test) {
+  test.equal(packageName, "core-files-next");
+});
+ */