+++ packages/core-files-next/index.ts
@@ -1,0 +1,10 @@
+// Write your package code here!
+
+// Variables exported by this module can be imported by other packages and
+// applications. See core-files-next-tests.js for an example of importing.
+
+export * from './plugins/minio';
+
+export const name = 'unchained:core-files-next';
+export * from './db';
+export default name;