+++ packages/core-quotations/db/quotation-documents/collections.js
@@ -1,7 +1,0 @@
-import { createFilesCollection } from 'meteor/unchained:core-files';
-
-export const QuotationDocuments = createFilesCollection('quotation_documents', {
-  maxSize: 10485760,
-});
-
-export default QuotationDocuments;