+++ packages/core-quotations/db/quotations/helpers.js
@@ -7,11 +7,16 @@ import { Countries } from 'meteor/unchained:core-countries';
 import { Currencies } from 'meteor/unchained:core-currencies';
 import { Logs, log } from 'meteor/unchained:core-logger';
 import { WorkerDirector } from 'meteor/unchained:core-worker';
+import {
+  uploadObjectStream,
+  uploadFileFromURL,
+  MediaObjects,
+} from 'meteor/unchained:core-files-next';
 import { Quotations } from './collections';
-import { QuotationDocuments } from '../quotation-documents/collections';
 import { QuotationStatus } from './schema';
 import { QuotationDirector } from '../../director';
 import settings from '../../settings';
+import { updateQuotationDocuments } from '../quotation-documents/helpers';
 
 Logs.helpers({
   quotation() {
@@ -188,28 +193,25 @@ Quotations.helpers({
   addDocument(objOrString, meta, options = {}) {
     if (typeof objOrString === 'string' || objOrString instanceof String) {
       return Promise.await(
-        QuotationDocuments.insertWithRemoteURL({
-          url: objOrString,
-          ...options,
-          meta: {
-            quotationId: this._id,
-            ...meta,
-          },
-        })
+        uploadFileFromURL(
+          'quotation-documents',
+          { fileLink: objOrString },
+          {
+            ...options,
+            meta: {
+              quotationId: this._id,
+              ...meta,
+            },
+          }
+        )
       );
     }
     const { rawFile, userId } = objOrString;
-    return Promise.await(
-      QuotationDocuments.insertWithRemoteBuffer({
-        file: rawFile,
-        userId,
-        ...options,
-        meta: {
-          quotationId: this._id,
-          ...meta,
-        },
-      })
-    );
+    return uploadObjectStream('quotation-documents', rawFile, {
+      userId,
+      quotationId: this._id,
+      ...meta,
+    });
   },
   documents(options) {
     const { type } = options || {};
@@ -217,9 +219,9 @@ Quotations.helpers({
     if (type) {
       selector['meta.type'] = type;
     }
-    return QuotationDocuments.find(selector, {
+    return MediaObjects.find(selector, {
       sort: { 'meta.date': -1 },
-    }).each();
+    }).fetch();
   },
   document(options) {
     const { type } = options || {};
@@ -227,7 +229,7 @@ Quotations.helpers({
     if (type) {
       selector['meta.type'] = type;
     }
-    return QuotationDocuments.findOne(selector, { sort: { 'meta.date': -1 } });
+    return MediaObjects.findOne(selector, { sort: { 'meta.date': -1 } });
   },
   logs({ limit, offset }) {
     const selector = { 'meta.quotationId': this._id };
@@ -361,7 +363,7 @@ Quotations.updateStatus = ({ status, quotationId, info = '' }) => {
     try {
       // we are now allowed to stop this process, else we could
       // end up with non-proposed but charged orders.
-      QuotationDocuments.updateDocuments({
+      updateQuotationDocuments({
         quotationId,
         date,
         ...modifier.$set,